requirements.txt
fastapi
uvicorn[standard]
pandas
langchain
langchain-openai
requests
pydantic
python-dotenv
